Fix HomieSetting template type errors: change int32_t to long
- Change HomieSetting<int32_t> to HomieSetting<long> in PoolController.hpp
- Update validator lambda parameter from int32_t to long in PoolController.cpp
- Remove duplicate setupProxy() declarations in PoolController.hpp
- Remove duplicate friend declaration in PoolController.hpp

Homie library v3.0.1 only provides template specializations for long, bool,
double, and const char*, not for int or int32_t. Using int32_t causes linker
errors: "undefined reference to HomieSetting<int>::HomieSetting(...)"

Fixes linker errors:
- undefined reference to _ZN12HomieSettingIiEC1EPKcS2_
- undefined reference to _ZNK12HomieSettingIiE3getEv
- undefined reference to _ZN12HomieSettingIiE15setDefaultValueEi
- undefined reference to _ZN12HomieSettingIiE12setValidatorE...
- undefined reference to vtable methods

Build should now link successfully for both ESP32 and ESP8266.

+++ src/PoolController.cpp
@@ -324,7 +324,7 @@ auto PoolControllerContext::setup() -> void {
   Homie_setBrand("smart-swimmingpool");
 
   // default interval of sending Temperature values
-  this->loopIntervalSetting_.setDefaultValue(TEMP_READ_INTERVAL).setValidator([](const int32_t candidate) -> bool {
+  this->loopIntervalSetting_.setDefaultValue(TEMP_READ_INTERVAL).setValidator([](const long candidate) -> bool {
     return candidate >= 0 && candidate <= 300;
   });
